requirements.txt
typer
openai
python-dotenv
rich
google-generativeai
anthropic
